+++ tests/03_impl_from.rs
@@ -6,9 +6,6 @@ nodyn::nodyn! {
     u32,
     [u8;4],
   }
-
-  impl From
-
 }
 
 fn main() {