+++ tests/07_trait_blocks.rs
@@ -7,8 +7,6 @@ nodyn::nodyn! {
         Xee<'b>,
   }
 
-  impl From
-
   impl {
         #[must_use]
         pub const fn foo(&self) -> &'static str {