+++ tests/10_advanced_example.rs
@@ -34,8 +34,6 @@ nodyn::nodyn! {
         JsonArray,
     }
 
-    impl From
-
     impl fmt::Display {
         fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result;
     }